Add multidictionary (multi-lang) support to zxcvbn
Dictionary files can be added to dictionaries-available
Dictionaries you want to use should be linked from dictionaries-enabled.

Changed compile scripts to look for dictionaries, changed a bit how dictionary
object is build.

Changed the way zxcvbn-async is loading zxcvbn. Now it's loading it from the
same server (tries to guess it's path based on it's own) instead of loading it
from dropbox.com