Detect repeated matches and use entropy 1
Repeating a term should not increase the entropy as much as including an entirely new term.

Detect repetition of terms when calculating the score and use an entropy of 1 for every repeated term. This entropy basically represents the choice to repeat the term or not.

This also detects the repetition of subterms. For example 'asdfasdr' will detect the repetition of the word 'as' even though the match 'asdf' is used for calculating the entropy of the initial four letters. I consider this a feature but it's probably open for discussion.

Repetition of bruteforce matches can not be detected so far. This is really a pitty as it means passwords like ' . . . . .' still get a high entropy.

Having bruteforce matches in the list of matches might solve this. But might have other issues.