Update count_wikipedia.py for Python 3 compatibility and parallel execution
This commit guarantee full Python 2 and Python 3 compatibility.
It also massively improves the performance on multi-processor machines,
by using multiprossing.Pool.imap() to delegate the file reading and
string processing to sub-processes.

Only drawback in the current version is the missing of the periodic
logging of the currently processed file.