Update README.md with updated C# port link
The previous C# port link was directed to a 404'd page. The proposed replacement link is for a project that appears to be the most up-to-date and has a dot.net NuGet available. The current revision of the proposed C# port is 7.0.92. Other ports seem to remain at 1.#.# and have less activity on their forks.